Implemented bow and arrow mechanics
Player can now equip a bow and arrow item. They can hold the action item to ready the bow and release to loose the arrow. The arrow will travel in a line, damaging any enemies it contacts. When it strikes a wall it plays a brief "arrow sticking" animation before disappearing.