Added new starting area and sword rooms
Re-defined base room from which all other rooms are derived for better organization of UI, Effects, Instances, Collision, Tilesets, and Background information and assets. Created two child rooms to the base: the starting area and the room in which the player can pick up the sword as their first item.